cleaning up MapFilterContextProvider
